feat: containerize supply chain platform

+++ app/models/supplier.py
@@ -1,5 +1,6 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, Numeric
 from app.database import Base
+
 
 class Supplier(Base):
     __tablename__ = "suppliers"
@@ -8,4 +9,4 @@ class Supplier(Base):
     name = Column(String, nullable=False)
     country = Column(String, nullable=False)
     risk_level = Column(String, nullable=False)
-    blocked_stock_eur = Column(Integer, nullable=False)
+    blocked_stock_eur = Column(Numeric(12, 2), nullable=False)
